Classifying image as "Day" or "Night" using Python.
Here's an algorithm that takes a .jpg image returns a string with either of the two values: “Day” or “Night”. It includes PIL to read the image, and Numpy to do data processing. 

The code can be viewed in the "Day_Night_Classifier.py" file and the sample input images are stored in "sample" folder.